clasifier/requirements.txt
tensorflow
tensorflow-io
matplotlib
kafka-python
